requirements.txt
Pillow>=9.0.0
pyembroidery>=1.4.0
Flask>=2.3.0
numpy>=1.24.0